test(doctor): snapshot redirected TUI profile
