+++ crates/vize/src/commands/doctor/tui/tests.rs
@@ -1,48 +1,21 @@
+mod snapshot_tests;
+
 use std::path::PathBuf;
 
-use vize_carton::String;
 use vize_doctor::{
     AnalysisProvenance, DoctorCategory, DoctorFinding, DoctorReport, EvidenceKind,
     FindingAssessment, FindingConfidence, FindingEvidence, FindingImpact, FindingSeverity,
     HealthPenalty, RuleCost, SourceLocation,
 };
 use vize_fresco::{
-    Buffer, ColorPreference, DiagnosticWorkspaceFocus, DiagnosticWorkspaceMode,
-    DiagnosticWorkspacePane, FeaturePreference, Key, KeyEvent, TerminalCapabilities,
+    ColorPreference, FeaturePreference, Key, KeyEvent, TerminalCapabilities,
     TerminalCapabilityProbe, TerminalProfileOptions,
 };
 
 use super::{
     DoctorSource, editor_command,
     model::{DoctorTuiModel, InteractionMode, InteractionOutcome},
-    render::render_frame,
 };
-
-#[test]
-fn frame_contains_semantic_summary_list_detail_and_evidence() {
-    let report = report();
-    let sources = sources();
-    let mut model = DoctorTuiModel::new(&report, 100, 18);
-    let mut buffer = Buffer::new(100, 18);
-
-    render_frame(
-        &mut buffer,
-        &mut model,
-        &sources,
-        capabilities(100, 18, true),
-    )
-    .unwrap();
-    let screen = screen_text(&buffer);
-
-    assert!(screen.contains("VIZE DOCTOR"));
-    assert!(screen.contains("Score: 95 / 100"));
-    assert!(screen.contains("VIZE_TEST_ERROR"));
-    assert!(screen.contains("Severity: error"));
-    assert!(screen.contains("Evidence"));
-    assert!(screen.contains("component: Parent passes mutable state"));
-    assert!(screen.contains("Fix safety: unavailable"));
-    insta::assert_snapshot!("doctor_tui_wide", screen);
-}
 
 #[test]
 fn navigation_filters_and_incremental_search_preserve_stable_selection() {
@@ -81,74 +54,6 @@ fn navigation_filters_and_incremental_search_preserve_stable_selection() {
         InteractionOutcome::Changed
     );
     assert_eq!(search.mode(), InteractionMode::Browse);
-}
-
-#[test]
-fn evidence_focus_and_narrow_stacked_navigation_use_fresco_state() {
-    let report = report();
-    let mut model = DoctorTuiModel::new(&report, 100, 18);
-
-    assert_eq!(
-        model.handle_key(&KeyEvent::char(']')),
-        InteractionOutcome::Changed
-    );
-    assert_eq!(model.selected_evidence_key(), Some(1));
-    assert_eq!(
-        model.workspace().focus(),
-        DiagnosticWorkspaceFocus::Evidence
-    );
-
-    model.resize(50, 14);
-    assert_eq!(
-        model.workspace().layout().mode(),
-        DiagnosticWorkspaceMode::Stacked
-    );
-    let _ = model.workspace().active_stacked_pane();
-    assert_eq!(
-        model.workspace().active_stacked_pane(),
-        DiagnosticWorkspacePane::Detail
-    );
-    assert_eq!(
-        model.handle_key(&KeyEvent::key(Key::Tab)),
-        InteractionOutcome::Changed
-    );
-    assert_eq!(
-        model.workspace().active_stacked_pane(),
-        DiagnosticWorkspacePane::Findings
-    );
-
-    let mut buffer = Buffer::new(50, 14);
-    render_frame(
-        &mut buffer,
-        &mut model,
-        &sources(),
-        capabilities(50, 14, true),
-    )
-    .unwrap();
-    insta::assert_snapshot!("doctor_tui_narrow", screen_text(&buffer));
-}
-
-#[test]
-fn ascii_monochrome_profile_retains_every_meaning_without_color() {
-    let report = report();
-    let sources = sources();
-    let mut model = DoctorTuiModel::new(&report, 100, 16);
-    let mut buffer = Buffer::new(100, 16);
-
-    render_frame(
-        &mut buffer,
-        &mut model,
-        &sources,
-        capabilities(100, 16, false),
-    )
-    .unwrap();
-    let screen = screen_text(&buffer);
-
-    assert!(screen.contains("x Severity: error"));
-    assert!(screen.contains("x Impact: high"));
-    assert!(!screen.contains('│'));
-    assert!(buffer.iter().all(|(_, _, cell)| cell.style.fg.is_none()));
-    insta::assert_snapshot!("doctor_tui_ascii_monochrome", screen);
 }
 
 #[test]
@@ -316,22 +221,4 @@ fn capabilities(width: u16, height: u16, unicode: bool) -> TerminalCapabilities 
             narrow_width: 60,
         },
     )
-}
-
-fn screen_text(buffer: &Buffer) -> String {
-    let mut screen = String::new("");
-    for y in 0..buffer.height() {
-        if y > 0 {
-            screen.push('\n');
-        }
-        let mut line = String::new("");
-        for x in 0..buffer.width() {
-            let cell = buffer.get(x, y).unwrap();
-            if !cell.is_continuation {
-                line.push_str(&cell.symbol);
-            }
-        }
-        screen.push_str(line.trim_end());
-    }
-    screen
 }
